# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CodeFabric Admin API",
        "description": "CodeFabric 管理后台后端接口文档（Laravel 13）",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "本地开发环境"
        }
    ],
    "paths": {
...
        "/api/admin/operation-logs": {
            "get": {
                "tags": [
                    "操作日志"
                ],
                "description": "分页查询操作日志",
                "operationId": "adminOperationLogList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页码，从 1 开始",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "maximum": 200,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "path",
                        "in": "query",
                        "description": "请求路径关键字",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "module",
                        "in": "query",
                        "description": "模块名称关键字",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 100
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "用户名关键字",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 100
                        }
                    },
                    {
                        "name": "success",
                        "in": "query",
                        "description": "是否成功（true/false/1/0）",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "true",
                                "false",
                                "1",
                                "0"
                            ]
                        }
                    },
                    {
                        "name": "startTime",
                        "in": "query",
                        "description": "开始时间",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "endTime",
                        "in": "query",
                        "description": "结束时间",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OperationLogListEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "操作日志"
                ],
                "description": "清理操作日志，不传 ids 时清空全部",
                "operationId": "adminOperationLogClear",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClearByIdsRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "清理成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "参数校验失败",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/operation-logs/{id}": {
            "get": {
                "tags": [
                    "操作日志"
                ],
                "description": "获取操作日志详情",
                "operationId": "adminOperationLogDetail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "日志 ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OperationLogDetailEnvelope"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "操作日志不存在",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    },
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClearByIdsRequest": {
                "properties": {
                    "ids": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "minimum": 1
                        },
                        "example": [
                            1,
                            2,
                            3
                        ]
                    }
                },
                "type": "object"
            },
...
            "OperationLogListItem": {
                "required": [
                    "id",
                    "path",
                    "method",
                    "statusCode",
                    "success",
                    "createdAt"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2001
                    },
                    "module": {
                        "type": "string",
                        "example": "用户管理",
                        "nullable": true
                    },
                    "action": {
                        "type": "string",
                        "example": "编辑用户",
                        "nullable": true
                    },
                    "username": {
                        "type": "string",
                        "example": "admin",
                        "nullable": true
                    },
                    "ip": {
                        "type": "string",
                        "example": "127.0.0.1",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "example": "未知地点",
                        "nullable": true
                    },
                    "path": {
                        "type": "string",
                        "example": "/api/admin/users/1"
                    },
                    "method": {
                        "type": "string",
                        "example": "PUT"
                    },
                    "statusCode": {
                        "type": "integer",
                        "example": 200
                    },
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "ok",
                        "nullable": true
                    },
                    "durationMs": {
                        "type": "integer",
                        "example": 38,
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "example": "2026-04-05 10:00:00"
                    }
                },
                "type": "object"
            },
            "OperationLogPageData": {
                "required": [
                    "list",
                    "total",
                    "page",
                    "pageSize"
                ],
                "properties": {
                    "list": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OperationLogListItem"
                        }
                    },
                    "total": {
                        "type": "integer",
                        "example": 256
                    },
                    "page": {
                        "type": "integer",
                        "example": 1
                    },
                    "pageSize": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "OperationLogListEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "$ref": "#/components/schemas/OperationLogPageData"
                    }
                },
                "type": "object"
            },
            "OperationLogDetailItem": {
                "required": [
                    "id",
                    "path",
                    "method",
                    "statusCode",
                    "success",
                    "createdAt"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 2001
                    },
                    "userId": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "module": {
                        "type": "string",
                        "example": "用户管理",
                        "nullable": true
                    },
                    "action": {
                        "type": "string",
                        "example": "编辑用户",
                        "nullable": true
                    },
                    "username": {
                        "type": "string",
                        "example": "admin",
                        "nullable": true
                    },
                    "ip": {
                        "type": "string",
                        "example": "127.0.0.1",
                        "nullable": true
                    },
                    "location": {
                        "type": "string",
                        "example": "未知地点",
                        "nullable": true
                    },
                    "path": {
                        "type": "string",
                        "example": "/api/admin/users/1"
                    },
                    "method": {
                        "type": "string",
                        "example": "PUT"
                    },
                    "statusCode": {
                        "type": "integer",
                        "example": 200
                    },
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "ok",
                        "nullable": true
                    },
                    "durationMs": {
                        "type": "integer",
                        "example": 38,
                        "nullable": true
                    },
                    "createdAt": {
                        "type": "string",
                        "example": "2026-04-05 10:00:00"
                    },
                    "userAgent": {
                        "type": "string",
                        "example": "Mozilla/5.0",
                        "nullable": true
                    },
                    "requestBody": {
                        "nullable": true
                    },
                    "responseBody": {
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "OperationLogDetailEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "$ref": "#/components/schemas/OperationLogDetailItem"
                    }
                },
                "type": "object"
            },
...
            "ApiEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "description": "CodeFabric 后端 OpenAPI 根定义。\n\n说明：\n- 使用 swagger-php Attributes 维护接口文档。\n- 业务返回统一为 { code, message, data }，通过组件 Schema 复用。",
                        "type": "integer",
                        "example": 0
                    },
                    "message": {
                        "type": "string",
                        "example": "ok"
                    },
                    "data": {
                        "nullable": true,
                        "oneOf": [
                            {
                                "type": "object",
                                "additionalProperties": true
                            },
                            {
                                "type": "array",
                                "items": {
                                    "type": "object",
                                    "additionalProperties": true
                                }
                            },
                            {
                                "type": "string"
                            },
                            {
                                "type": "integer"
                            },
                            {
                                "type": "boolean"
                            }
                        ]
                    }
                },
                "type": "object"
            },
            "ApiErrorEnvelope": {
                "required": [
                    "code",
                    "message",
                    "data"
                ],
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 2004
                    },
                    "message": {
                        "type": "string",
                        "example": "请求失败"
                    },
                    "data": {
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "在 Authorization 请求头中传入 Bearer Token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "cookieAuth": {
                "type": "apiKey",
                "description": "浏览器 Cookie 鉴权",
                "name": "admin_token",
                "in": "cookie"
            }
        }
    },
...
}